Move `ReplaceFileSource` command
