+++ wcfsetup/install/files/lib/system/file/processor/FileProcessor.class.php
@@ -13,7 +13,7 @@ use wcf\event\file\GenerateWebpVariant;
 use wcf\system\database\util\PreparedStatementConditionBuilder;
 use wcf\system\event\EventHandler;
 use wcf\system\exception\SystemException;
-use wcf\system\file\command\ReplaceFileSource;
+use wcf\command\file\ReplaceFileSource;
 use wcf\command\file\ReplaceWithWebpVariant;
 use wcf\system\file\processor\exception\DamagedImage;
 use wcf\system\image\adapter\exception\ImageNotProcessable;
